Ability to compute the nodule tri-dimensional measure and report bounds
We add the ability to compute the tri-dimensional measure. This involves 3 linear measures which are all perpendicular to one another. The first measure is RECIST, which is the longest in-plane length of the nodule. The second is another inplane measure which is perpendicular to the RECIST and measures the largest distance between two points on the nodule surface on the same axial plane as RECIST, while peing perpendicular to the RECIST measure. The third is a measure that is perpendicular to both the aforementioned diameters and passes through their intersection point and measures the widest Z span of the nodule.

We also report these numbers. We also report the nodule span along the X,Y and Z axes